Refactor logging middleware to replace `StructuredLogger` with `WebRequestLogger` and improve request logging
- Replaced `StructuredLogger` middleware with `WebRequestLogger` for consistent structured logging of web requests.
- Updated router setup and associated tests to use `WebRequestLogger` middleware.
- Added `RequestIDLogger` middleware to enhance request tracking capabilities.
