+++ internal/web_request_logger.go
@@ -1,0 +1,48 @@
+package internal
+
+import (
+	"log/slog"
+	"net/http"
+	"time"
+)
+
+// WebRequestLogger returns a middleware that logs each request in structured (JSON) form using slog.
+func WebRequestLogger(next http.Handler) http.Handler {
+
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		start := time.Now()
+
+		// Use a wrapper to get the status code
+		ww := &responseWriter{ResponseWriter: w, status: http.StatusOK}
+		next.ServeHTTP(ww, r)
+
+		duration := time.Since(start)
+
+		logger := LoggerFromContext(r.Context())
+		//requestId := GetRequestIdFromContext(r.Context())
+
+		logger.Info("WEB_REQUEST",
+			slog.String("method", r.Method),
+			slog.String("path", r.URL.Path),
+			slog.Int("status", ww.status),
+			slog.String("remote", r.RemoteAddr),
+			slog.String("user_agent", r.UserAgent()),
+			slog.Int64("duration_ms", duration.Milliseconds()),
+			//slog.String("request_id", requestId),
+		)
+	})
+}
+
+// responseWriter wraps http.ResponseWriter to capture the status code
+type responseWriter struct {
+	http.ResponseWriter
+	status int
+}
+
+func (rw *responseWriter) WriteHeader(code int) {
+	rw.status = code
+	rw.ResponseWriter.WriteHeader(code)
+}
+func (rw *responseWriter) Write(b []byte) (int, error) {
+	return rw.ResponseWriter.Write(b)
+}
